Support non-integer speed values with 0.5 step increments
- Change speed slider step from 1 to 0.5 for finer control
- Use parseFloat instead of parseInt in JS for speed values
- Backend accepts float speed indices, interpolates between
  adjacent speed/duration entries for in-between values
- inc_speed/dec_speed now step by 0.5

+++ src/cube/presentation/gui/backends/webgl/ClientSession.py
@@ -363,9 +363,9 @@ class ClientSession:
         if command:
             self.inject_command(command)
 
-    def _handle_speed(self, speed_index: int) -> None:
+    def _handle_speed(self, speed_index: float) -> None:
         from cube.application.state import speeds
-        clamped = max(0, min(len(speeds) - 1, speed_index))
+        clamped = max(0.0, min(len(speeds) - 1, float(speed_index)))
         self._app.vs._speed = clamped
 
     def _handle_size(self, size: int) -> None:
